[hotfix] Rename CloneHiveUtilsTest to ListHiveFilesFunctionTest
